refactor(deps): migrate all require() call sites to extra-name guard (#322)
