+++ src/autointent/modules/scoring/_lora/lora.py
@@ -6,8 +6,8 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal
 
 from autointent import Context
+from autointent._deps import require
 from autointent._dump_tools import Dumper
-from autointent._utils import require
 from autointent.configs import EarlyStoppingConfig, HFModelConfig
 from autointent.modules.scoring._bert import BertScorer
 
@@ -75,7 +75,7 @@ class BERTLoRAScorer(BertScorer):
         **lora_kwargs: Any,  # noqa: ANN401
     ) -> None:
         # Lazy import peft
-        require("peft", extra="peft")
+        require("peft")
         from peft import LoraConfig
 
         # early stopping doesn't work with lora for now https://github.com/huggingface/transformers/issues/38130
